Implement unit tests for MLAKVCache, DeepSeekV3Model, and various components
- Added unit tests for MLAKVCache to verify update, overflow rejection, and reset functionality.
- Created end-to-end tests for DeepSeekV3Model to ensure correct behavior in training and evaluation modes.
- Introduced tests for loss computation, learning rate scheduling, and optimizer functionality.
- Implemented tests for MoE layer and its routing logic, ensuring correct token distribution and shape preservation.
- Added distributed utility tests for gradient clipping and tensor reduction operations.
- Included tests for rotary embeddings to validate frequency precomputation and application.
- Enhanced test coverage for parallel dimensions and their mesh construction in distributed settings.
- Updated the ScaledDotProductAttentionWrapper to support causal masking with KV-cache decoding.
- Added pytest as a development dependency for testing framework.

+++ src/model/moe/moe.py
@@ -198,8 +198,11 @@ class TokenChoiceTopKRouter(nn.Module):
         top_scores = top_scores * self.route_scale
 
         # num_tokens_per_expert: (num_experts,)
+        # NOTE: torch.histc's CPU kernel only supports floating-point inputs (CUDA supports
+        # integer types), so cast the (small, exact-in-float32) expert indices to float32
+        # before histogramming to keep this working identically on CPU and CUDA.
         num_tokens_per_expert = torch.histc(
-            selected_experts_indices.view(-1),
+            selected_experts_indices.view(-1).float(),
             bins=self.num_experts,
             min=0,
             max=self.num_experts,
@@ -250,8 +253,11 @@ class TokenReorderer(nn.Module):
         # num_tokens_per_expert = tensor([
         #     2., 3., 3., 2., 3., 3., 3., 4., 3., 2., 2., 2.
         # ])
+        # NOTE: torch.histc's CPU kernel only supports floating-point inputs (CUDA supports
+        # integer types), so cast the (small, exact-in-float32) expert indices to float32
+        # before histogramming to keep this working identically on CPU and CUDA.
         num_tokens_per_expert = torch.histc(
-            selected_experts_indices.view(-1),
+            selected_experts_indices.view(-1).float(),
             bins=self.num_experts,
             min=0,
             max=self.num_experts,
